+++ mathics/eval/associations/__init__.py
@@ -1,0 +1,4 @@
+"""
+Evaluation routines and associated code for Built-in functions found under module
+mathics.builtins.assications.
+"""
